requirements.txt
elasticsearch==9.0.1
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
kagglehub==0.3.12
ollama==0.4.4
black==24.4.2
